Modify function documentation a bit
The title and description don't need specific @title and @description roxygen tags so I removed those.

I did a general word usage and style pass to make sure proper terminology is used (e.g., package => Data Package)

It's a convention in the R community for help pages not to have code in the description so I tweaked the style a bit to avoid this. e.g., check_download_size = TRUE to \code{check_download_size} to \code{TRUE} .

Used \code{} markup so its clear which text is a function name or R expression.