Tweak how download_directory works
Since it's common in R to have downloading/saving functions just save to the current working directory, I think it's good to just make this the default. In general, I prefer to let the user run my functions with as little specification as possible.

Second thing I fixed is that I use file.path instead of paste. Not all operating systems use / as a path seperator so it's better to use file.path to construct an appropriate file path for the user's operating system