Remove @imports and add helpful check for pbapply package
Since we're only using a few functions from the dataone and pbapply package, it's generally encouraged to just explicitly call the functions with the :: syntax, rather than import all the names (functions) from both packages into the arcticdatautils namespace.

I also added a stop condition at the top of the function so the function errors out if the pbapply package isn't installed. Since arcticdatautils doesn't rely on pbapply, we would consider pbapply an optional dependency. So we'd add a if (!requireNamespace(..)) call like I did and then add it to the DESCRIPTION file at the top level of the package in the Suggests section. See http://r-pkgs.had.co.nz/description.html for more on this.