Use seq_len instead of 1:{x}
While it's commonly taught to new R programmers to use the sequence shorthand ":" to generate numeric vectors to use in for loops, this is error prone when the "to" side is not a number equal to or greater than the "from" side (e.g., 1:0). Using : is often safe, but using seq_len(n) is safer across more types of input. A common example to show this is when we want to iterate over the rows of an empty data.frame:

mydf <- data.frame()
1:nrow(mydf)

which fails because there isn't a 0th row.