Calculate total package size from the Solr index
While it's totally valid to use getSystemMetadat to get the file size of each Object we're about to download, the Solr index should already contain this information and we can easily query for it in one go. I wrote a re-usable function to do this, then modified your code to make use of it.