Change allocation of empty vector from c() to vector(...)
This is something that's really weird about R. When you create an empty vector with the c() function, it's the same as saying myvar <- NULL which is not really what you mean. This works out in many cases but some functions do weird things. In our case here, this was making an sapply call below return a list instead of a character vector which broke things.