Simplify the block of code to get child data PIDs
My code is essentially equivalent but there's something here that I like (hint: this is subjective). I'd rather just smush data_pids and child_data_pids together, not caring if they are empty or not. This is opposed to having to write an if statement to check if the variable was defined or not.